Call pivot_root to set container's root mount
fixes #435

`pivot_root` takes 2 arguments: `new_root` and `old_root`. In case of
success, `new_root` is set as the root mount for the current and new
processes and threads. Due to the restrictions described in
`pivot_root(2)`, `old_root` has to be underneath `new_root`. This means,
that we need a mounting point inside the `rootfs` of every container.
This patch adds `/mnt` to the container image packed with `squashfs` to
be used as the `old_root` parameter.